+++ openwhisk/extractor.go
@@ -18,65 +18,11 @@
 package openwhisk
 
 import (
-	"archive/zip"
-	"bytes"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
-	"path/filepath"
 	"strconv"
 )
-
-func unzip(src []byte, dest string) error {
-	reader := bytes.NewReader(src)
-	r, err := zip.NewReader(reader, int64(len(src)))
-	if err != nil {
-		return err
-	}
-
-	os.MkdirAll(dest, 0755)
-
-	// Closure to address file descriptors issue with all the deferred .Close() methods
-	extractAndWriteFile := func(f *zip.File) error {
-		rc, err := f.Open()
-		if err != nil {
-			return err
-		}
-		defer func() {
-			rc.Close()
-		}()
-
-		path := filepath.Join(dest, f.Name)
-
-		if f.FileInfo().IsDir() {
-			os.MkdirAll(path, f.Mode())
-		} else {
-			os.MkdirAll(filepath.Dir(path), f.Mode())
-			f, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
-			if err != nil {
-				return err
-			}
-			defer func() {
-				f.Close()
-			}()
-
-			_, err = io.Copy(f, rc)
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-	}
-
-	for _, f := range r.File {
-		err := extractAndWriteFile(f)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
 
 // higherDir will find the highest numeric name a sub directory has
 // 0 if no numeric dir names found
@@ -111,7 +57,7 @@ func (ap *ActionProxy) ExtractAction(buf *[]byte, suffix string) (string, error)
 	file := newDir + "/exec"
 	if IsZip(*buf) {
 		Debug("Extract Action, assuming a zip")
-		return file, unzip(*buf, newDir)
+		return file, Unzip(*buf, newDir)
 	}
 	return file, ioutil.WriteFile(file, *buf, 0755)
 }
